Add files via upload
I have been the main director and team action manager for the analysis and gathering of insight for an ideal possible improvement of a Maltese Company named QGen. The purpose of the project was to analyze how this organization can leverage emerging technologies to improve its business activities by adopting a sociotechnical system perspective. We as a team examined the organization's operations, assessed potential impact of emerging technologies in the operations, and proposed management strategies for implementing these technologies.